Create a wrapper for CarStatsLog.
This allows us to avoid use spyStatic on CarStatsLog. spyStatic
is not thread safe and if CarStatsLog is accessed while mockito
session is setting up, weird error may happen, which causes flaky
tests.

This CL creates a wrapper class that wraps the static method so
that we can use a per-test mocked class instead.

Test: atest CarServiceWatchdogUnitTest CarServiceHalUnitTest \
IoOveruseHandlerUnitTest
Flag: EXEMPT refactor
Bug: 409617758

Change-Id: Ie85cea980849051bcbc6e8df1933d56fb4816e2b