Map detailed not available status for legacy client.
If the client target sdk version is < 25Q4, we must map the new
detailed not available property status to the general not available
property status.

This CL also refactors CarPropertyValue so that it is still immutable.
We try to keep it immutable so that we do not need to worry about
thread safety.

We currently do not have a version code for 25Q4 yet, so we use
CUR_DEVELOPMENT in this CL. It must be updated later.

Test: atest CarPropertyManagerUnitTest \
CarPropertyEventCallbackControllerUnitTest
Flag: android.car.feature.car_property_status_detailed_not_available
Bug: 416768353

Change-Id: I13c5b40e2b7a238fcc3d057a06dd1f724ed764ae