Set defer parameter to false if policy already applied
In the case that a power policy change is requested by CPMS, CPPD can respond with onApplyPowerPolicySucceeded, which has a 'deferred' parameter. When the policy change request was successful, but the requested policy is already the current one, the notification of that policy change shouldn't be deferred, since it was already applied before.

Bug: 402686744
Test: atest carpowerpolicyserver_test CarPowerHostTest PowerPolicyHostTest CarPowerManagerTest
Flag: EXEMPT bugfix
Change-Id: I4aa757f71c52e6c5fde71701184be2cfc0a56ed6