Initialize mUserRemoveLatch for each test.
Previously the latch was initialized only once for the class, and one
test that removed a user decrementing the latch would affect other
tests. Now we initialize the latch at the begninning of each test, so
that each test in the class can run withoout affecting each other.

Bug: 398485376
Flag: TEST_ONLY
Test: atest CarHiddenApiTest:android.car.hiddenapitest.CarDevicePolicyManagerTest

Change-Id: Id8c22b81266c50cdd109cdc9aafd61123f81ab05