Replace call to quitSafely with call to releaseHandlerThread
When a caller of getHandlerThread calls quitSafely on its handler
thread, while the thread will quit, it will not remove that thread from
the cache located in CarServiceUtils. This can lead to a race condition
where the handler thread is used in future tests because it is still in
the cache. Call releaserHandlerThread instead of quitSafely to ensure
that the handler thread is removed from the cache.

Test: atest CarWatchdogDaemonHelperTest
Flag: EXEMPT refactor
Fixes: 416777246
Fixes: 422989055
Fixes: 430064298
Change-Id: I5d38d6a611c7b34a79e60a31e4534d1167a8c7df