Revert "Revert "Map detailed not available status for legacy cli..."
Revert submission 33928451-revert-33607460-VVDAOGWIKG

Reason for revert: All tests passing now. Changed 25Q4 refs to 26Q2.
Added missing call to map system status in onEvent which is why tests
were failing.

Reverted changes: /q/submissionid:33928451-revert-33607460-VVDAOGWIKG

Map detailed not available status for legacy client.

If the client target sdk version is < 26Q2, we must map the new
detailed not available property status to the general not available
property status.

This CL also refactors CarPropertyValue so that it is still immutable.
We try to keep it immutable so that we do not need to worry about
thread safety.

We currently do not have a version code for 26Q2 yet, so we use
CUR_DEVELOPMENT in this CL. It must be updated later.

Test: atest CarPropertyManagerUnitTest \
CarPropertyEventCallbackControllerUnitTest
Flag: android.car.feature.car_property_status_detailed_not_available
Bug: 416768353

Change-Id: I7f7563ad7c439ea4e1e90a0114ff8ff4cc42a0af