Add empty previously released test API
This is needed to make the following change (http://ag/34891836) work
correctly. This is similar to change https://r.android.com/3058122
which added empty system APIs. In neither cases was the previously
released API file an accurate representation of the APIs at the
specified version but trying to determine the exact API at the time
version 3 was dropped was extremely difficult.

Flag: EXEMPT build change to improve flag support
Bug: 425309457
Test: TH
Change-Id: Ic3051d2f271128b6733447d03d13b286927177f3