Set lock screen disabled for all created users
On Automotive form factors, there is only a lock screen present for
secure users and by default, users are not secure. Disabling the lock
screen by default will improve wake and boot performance by shortcutting
the keyguard flow for non-secure users.

This value does not require
updating because lock screen will always be shown for secure users
regardless of the setting of this value. The only valid use case for
having lock screen enabled without a secure lock is for the "swipe"
lockscreen which is not present on Automotive.

Bug: 430639961
Test: manual, CarUserServiceTest
Flag: NONE bugfix
Change-Id: Id3db0b3c6b556d961280a0910c57e51356e82611