Add Perfetto controller fragment to KitchenSink.
This commit introduces a new fragment to the EmbeddedKitchenSinkApp that
allows for manual control and testing of Perfetto tracing.

The new "Perfetto" screen provides buttons to:
- Push a default Perfetto trace configuration.
- Trigger a Perfetto trace.
- Query the currently active Perfetto configuration.
- Remove any existing Perfetto configurations.

A status view displays the results of these operations, and a
floating action button is included to scroll the view to the top.

This UI will aid in the development and debugging of Perfetto-based
tracing features.

Bug: 406520911
Test: Manually tested in EmbeddedKitchenSinkApp
Flag: EXEMPT TEST_ONLY

Change-Id: Ie0e7a70adcb0d512e3773d52366ce0665f83cb1d