Set home component states based on user rather than display
Depending on the timings of the user switch, the occupant zone service
may not have the up-to-date userId for the driver during the various
lifecycle events. By using the current user to check for driver, the
check is more consistent.

Bug: 441579704
Test: manual (checked home on md emulator across reboots/user switches)
atest CarUserServiceTest
Flag: android.car.feature.visible_background_user_toggle_home_components

Change-Id: I4655b3b24ccf1c7d9afd553baf2ef12137240d71