Pass PackageInfoResolver to objects to remove call to getInstance
libwatchdog_test crashes due to PackageInfoResolver::getInstance being
invoked in tests before a MockPackageInfoResolver is assigned, but
PackageInfoResolver::terminate is not called after. Pass an instance of
PackageInfoResolver to UidStatsCollector, WatchdogProcessService,
WatchdogPerfService, and IoOveruseMonitor as well as base or derived
classes to prevent calls to PackageInfoResolver::getInstance during
tests.

Test: atest libwatchdog_test
Test: atest libpilotwatchdog_test
Test: atest libwatchdogbase_test
Flag: EXEMPT TEST_ONLY
Fixes: 442904169
Fixes: 442747644
Change-Id: Ia0b2c16344060410c6707cd0c8dfff79f6d27648