Implement `getPhysicalCameraInfo` in `CompatVirtualCamera`.
This change enables `CompatVirtualCamera` to return the `CameraDesc` of a physical camera. The `CompatHalCamera` now stores the `CameraDesc` associated with the physical camera, which is passed from `CompatEnumerator` during initialization. `CompatEnumerator` also renames `mCameraDesc` to `mCameraDescs` for better readability.

The change is implicitly guarded by the flag, as CompatHalCamera and CompatVirtualCamera are instantiated by CompatEnumerator which is guarded by a flag.

Test: atest evs_compat_test
Bug: 441577862
Flag: android.car.feature.car_evs_compat_lib
Change-Id: I3fae00b163fe1477b13e73718c513cbeccdca73b