Implement NDK Camera stream management in CompatHalCamera.
This change adds the functionality to start and manage a camera stream using the Android NDK Camera API within the `CompatHalCamera`. It includes setting up an `AImageReader` to acquire frames, defining callbacks for camera session and capture events, and converting the received `AImage` objects into `BufferDesc` structures. The `clientStreamStarting` method is introduced to initiate the NDK stream setup.

The change is implicitly guarded by the flag, as CompatHalCamera and CompatVirtualCamera are instantiated by CompatEnumerator which is guarded by a flag.

Test: atest evs_compat_test
Bug: 441577862
Flag: android.car.feature.car_evs_compat_lib
Change-Id: I08bf479dc9896d982bc7c273cfca1338f751cb7d