Mock context for inhibit tests
Problem: The test creates an object under test with a user ID of 10. The
object then tries to get a context _as_ that user. Because the test
_also_ used a real context from the instrumentation registry, this
causes the system to check for an active foreground user 10. If the
tests ever end up running on a different user ID, they all fail.

Solution: Tests really shouldn't be using a real context object, as they
risk both bugs like this, and changing the actual system state with test
scenarios (like Settings Provider in this case). We should use a full
mock context and mock out the bits that we need.

Flag: EXEMPT TEST_ONLY
Bug: 453781914
Test: atest CarServiceBluetoothUnitTest
Change-Id: Ie3cc3438edd14998ff8ae39e3f207971db272706