Fix failed testSendEvent() on CW builds by handling null navigationBinder
Previously, `InstrumentClusterService.getInstrumentClusterInfo()` would throw an `IllegalStateException` if the `navigationBinder` remained null after multiple retries.

This change updates the behavior to return `null` instead of throwing an exception. This is because a null `navigationBinder` is an expected state when the `cluster_service` is not enabled, as `init()` would not have been called.

This modification prevents test failures in `android.car.cts.CarNavigationManagerTest#testSendEvent` on Automotive (CW) builds where the service might not be enabled.

Bug: 431497688
Test: atest CtsCarTestCases:android.car.cts.CarNavigationManagerTest#testSendEvent
Test: atest InstrumentClusterServiceTest
Flag: EXEMPT BUGFIX

Change-Id: Ia0edfc6b201f99e1112523720d99741fdc06a506