Add RootTaskListener to CarActivityService.
So that other services in CarService can listen to changes in root
tasks. For example ClusterHomeService wants to subscribe to changes in
root tasks, so that it can start the cluster activity in a root task and
stops it when the root task becomes unavailable.
Also implement CarActivityService#getRootTaskInfo method so root task
info can be queried by the root task name.

Bug: 437049703
Flag: android.car.feature.root_task_cluster
Test: atest CarActivityServiceUnitTest

Change-Id: Ia739a5eac9d4e0c2f4a162b21da05642611f7650