Fix event injection logic for test
Use the last event timestamp instead of the current elapsedRealtimenanos. The reason is that on certain devices such as CF, the VHAL properties are stored outside of Android and/or are using a different time base than Android. Therefore, injection may fail because the Android timestamp is older than the external timestamp.

Flag: EXEMPT TEST_ONLY
Bug: 459899699
Bug: 350822044
Test: atest com.android.car.vehiclehal.test.CarPropertyTest#testPropertyEventOutOfOrder
Change-Id: I11160d8e7066bd7ea964ad20b1a438296e926c57