Remove unused notification dimens (car)
Also updated the references to notification_content_margin_start to be
the actual value, since that dimen no longer exists in the platform. I'm
not sure what the intention was with the comments about
notification_content_plus_picture_margin, but that's out of scope for
this change since we want it to be a no-op.

Bug: 444294237
Flag: EXEMPT PURE_REFACTOR
Test: builds
Change-Id: Ida8b3eab145382300a07e3b1d26ec04c15b18f4d