Remove notification_template_right_icon
This is not used anywhere in the platform now. On handheld devices, this
layout was replaced with notification_2025_right_icon.xml, which will
soon be renamed to notification_right_icon.xml. If this override is
still relevant for car, it already wasn't working and it should be fixed
accordingly. For now I'm cleaning it up so that I can remove the
notification_content_margin_end dimen, which has been replaced by
notification_2025_margin (soon to be renamed notification_margin).

Bug: 444294237
Test: presubmit
Flag: EXEMPT PURE_REFACTOR
Change-Id: I09da68489181c66123b0e82e90ad3dc21a710a08